+++ src/Repository/EntityFactoryInterface.php
@@ -14,7 +14,7 @@ interface EntityFactoryInterface
      * @param array $document the raw document as found in the _source field of the raw Elasticsearch response
      * @param array $metaData contains all "underscore-fields" delivered in the raw Elasticsearch response (e.g. _score)
      *
-     * @return mixed
+     * @return object
      */
     public function fromDocument(array $document, array $metaData);
 }